Move optimize to matOptimize.
