--- src/matUtils/optimize.hpp
@@ -1,4 +1,0 @@
-#include "common.hpp"
-
-po::variables_map parse_optimize_command(po::parsed_options parsed);
-void optimize_main(po::parsed_options parsed);
